Change usage and invocations
Completely changed synopses and invocation of the tools.
And thus heavily reworked the internals of the tool.

It now always installs the associated server archive
when launching. And arguments can be added to the
executed java command. Versions are also parsed
to identify if a version type was specified.

The worlds are now also part of mcserver's data
and are transparently managed, without needing
user creations/deletions.

The progress bar of an archive download was
also reworked to adapt to the tty screen if
the columns width could be probed.

Also fixed an installation issue where
a the generated debian package would install
the manpage in the toplevel mandir directory.